Longitudinal Database of Texas School District Property Values and Tax Rates.

This dataset comes from the Texas Education Agency (TEA). It includes data recorded at the bottom of each ISD's Public Education Information Management System (PEIMS) Financial Actual Reports.

Usage

tea_isd_tax_data

Format

A tibble with 9,161 rows and 13 variables:

year

The year end of a school year. For example, 2013 refers to SY2012-2013.

dist_id

The District ID Assigned by the Texas Education Agency

dist_name

The District Name as Recognized by the Texas Education Agency

m_o_tax_rate

The Maintenance & Operations (M&O) Tax Rate for the year end recorded.

i_s_tax_rate

The Interest & Sinking (I&S) Tax Rate for the year end recorded.

total_tax_rate

The Total Tax Rate (M&O + I&S) for the year end recorded.

total_property_value

The year end's Certified Total Property Values Used by TEA

total_property_value_per_pupil

The year end's Certified Total Property Values Used by TEA, Per Pupil

business_property_value

The year end's Certified Business Property Values Used by TEA

residential_property_value

The year end's Certified Residential Property Values Used by TEA

land_property_value

The year end's Certified Land Property Values Used by TEA

oil_and_gas_property_value

The year end's Certified Oil + Gas Property Values Used by TEA

other_property_value

The year end's Certified Other Property Values Used by TEA

Source

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2012-2013: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2013-2014: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2014-2015: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2015-2016: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2016-2017: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2017-2018: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2018-2019: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2019-2020: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).

TEA's ISD Tax Rates and State Certified Property Values Data for All Texas School Districts, SY2020-2021: ISD Tax Rates and State Certified Property Values, via Texas Education Agency's Public Education Information Management System (PEIMS).